refactor(ethereum): collapse root fetcher
